tidy up UILineEdit

+++ include/LavenderUI/Core/EventHandler/IEvent.hpp
@@ -79,8 +79,8 @@ struct MouseDragEvt : public IEventCRTP<MouseDragEvt>
 struct MouseScrollEvt : public IEventCRTP<MouseScrollEvt>
 {
     MouseScrollEvt() {}
-    MouseScrollEvt(int8_t xOffsetIn, int8_t yOffsetIn) : xOffset{xOffsetIn}, yOffset{yOffsetIn} {}
-    int8_t xOffset{0}, yOffset{0};
+    MouseScrollEvt(int32_t xOffsetIn, int32_t yOffsetIn) : xOffset{xOffsetIn}, yOffset{yOffsetIn} {}
+    int32_t xOffset{0}, yOffset{0};
 };
 
 struct SliderEvt : public IEventCRTP<SliderEvt>
